# api/docs/swagger.yaml
basePath: /v1
definitions:
...
  entities.Phone:
    properties:
      created_at:
        example: "2022-06-05T14:26:02.302718+03:00"
        type: string
      fcm_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzd.....
        type: string
      id:
        example: 32343a19-da5e-4b1b-a767-3298a73703cb
        type: string
      max_send_attempts:
        description: MaxSendAttempts determines how many times to retry sending an
          SMS message
        example: 2
        type: integer
      message_expiration_seconds:
        description: MessageExpirationSeconds is the duration in seconds after sending
          a message when it is considered to be expired.
        type: integer
      message_send_schedule_id:
        example: 32343a19-da5e-4b1b-a767-3298a73703cb
        type: string
      messages_per_minute:
        example: 1
        type: integer
      missed_call_auto_reply:
        example: This phone cannot receive calls. Please send an SMS instead.
        type: string
      phone_number:
        example: "+18005550199"
        type: string
      sim:
        $ref: '#/definitions/entities.SIM'
      unarchive_thread:
        description: UnarchiveThread moves an archived message thread back to the
          inbox when a new message is received on this phone.
        example: false
        type: boolean
      updated_at:
        example: "2022-06-05T14:26:10.303278+03:00"
        type: string
      user_id:
        example: WB7DRDWrJZRGbYrv2CKGkqbzvqdC
        type: string
    required:
    - created_at
    - id
    - max_send_attempts
    - message_expiration_seconds
    - messages_per_minute
    - phone_number
    - sim
    - unarchive_thread
    - updated_at
    - user_id
    type: object
  entities.PhoneAPIKey:
    properties:
      api_key:
        example: pk_DGW8NwQp7mxKaSZ72Xq9v6xxxxx
        type: string
      created_at:
        example: "2022-06-05T14:26:02.302718+03:00"
        type: string
      id:
        example: 32343a19-da5e-4b1b-a767-3298a73703cb
        type: string
      name:
        example: Business Phone Key
        type: string
      phone_ids:
        example:
        - 32343a19-da5e-4b1b-a767-3298a73703cb
        - 32343a19-da5e-4b1b-a767-3298a73703cc
        items:
          type: string
        type: array
      phone_numbers:
        example:
        - "+18005550199"
        - "+18005550100"
        items:
          type: string
        type: array
      updated_at:
        example: "2022-06-05T14:26:02.302718+03:00"
        type: string
      user_email:
        example: user@gmail.com
        type: string
      user_id:
        example: WB7DRDWrJZRGbYrv2CKGkqbzvqdC
        type: string
    required:
    - api_key
    - created_at
    - id
    - name
    - phone_ids
    - phone_numbers
    - updated_at
    - user_email
    - user_id
    type: object
  entities.SIM:
    enum:
    - SIM1
    - SIM2
    type: string
    x-enum-varnames:
    - SIM1
    - SIM2
  entities.SubscriptionName:
    enum:
    - free
    - pro-monthly
    - pro-yearly
    - ultra-monthly
    - ultra-yearly
    - pro-lifetime
    - 20k-monthly
    - 100k-monthly
    - 50k-monthly
    - 200k-monthly
    - 20k-yearly
    type: string
    x-enum-varnames:
    - SubscriptionNameFree
    - SubscriptionNameProMonthly
    - SubscriptionNameProYearly
    - SubscriptionNameUltraMonthly
    - SubscriptionNameUltraYearly
    - SubscriptionNameProLifetime
    - SubscriptionName20KMonthly
    - SubscriptionName100KMonthly
    - SubscriptionName50KMonthly
    - SubscriptionName200KMonthly
    - SubscriptionName20KYearly
  entities.User:
    properties:
      active_phone_id:
        example: 32343a19-da5e-4b1b-a767-3298a73703cb
        type: string
      api_key:
        example: x-api-key
        type: string
      created_at:
        example: "2022-06-05T14:26:02.302718+03:00"
        type: string
      email:
        example: name@email.com
        type: string
      id:
        example: WB7DRDWrJZRGbYrv2CKGkqbzvqdC
        type: string
      notification_heartbeat_enabled:
        example: true
        type: boolean
      notification_message_status_enabled:
        example: true
        type: boolean
      notification_newsletter_enabled:
        example: true
        type: boolean
      notification_webhook_enabled:
        example: true
        type: boolean
      subscription_ends_at:
        example: "2022-06-05T14:26:02.302718+03:00"
        type: string
      subscription_id:
        example: 8f9c71b8-b84e-4417-8408-a62274f65a08
        type: string
      subscription_name:
        allOf:
        - $ref: '#/definitions/entities.SubscriptionName'
        example: free
      subscription_renews_at:
        example: "2022-06-05T14:26:02.302718+03:00"
        type: string
      subscription_status:
        example: on_trial
        type: string
      timezone:
        example: Europe/Helsinki
        type: string
      updated_at:
        example: "2022-06-05T14:26:10.303278+03:00"
        type: string
    required:
    - api_key
    - created_at
    - email
    - id
    - notification_heartbeat_enabled
    - notification_message_status_enabled
    - notification_newsletter_enabled
    - notification_webhook_enabled
    - subscription_id
    - subscription_name
    - timezone
    - updated_at
    type: object
...
  requests.PhoneAPIKeyStoreRequest:
    properties:
      name:
        example: My Phone API Key
        type: string
    required:
    - name
    type: object
...
  requests.PhoneUpsert:
    properties:
      fcm_token:
        description: FcmToken is either a Firebase registration token or a public
          HTTPS adapter callback URL.
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzd.....
        type: string
      max_send_attempts:
        description: MaxSendAttempts is the number of attempts when sending an SMS
          message to handle the case where the phone is offline.
        example: 2
        type: integer
      message_expiration_seconds:
        description: MessageExpirationSeconds is the duration in seconds after sending
          a message when it is considered to be expired.
        example: 12345
        type: integer
      message_send_schedule_id:
        example: 32343a19-da5e-4b1b-a767-3298a73703cb
        type: string
      messages_per_minute:
        example: 1
        type: integer
      missed_call_auto_reply:
        example: e.g. This phone cannot receive calls. Please send an SMS instead.
        type: string
      phone_number:
        example: "+18005550199"
        type: string
      sim:
        description: SIM is the SIM slot of the phone in case the phone has more than
          1 SIM slot
        example: SIM1
        type: string
      unarchive_thread:
        description: UnarchiveThread moves an archived thread back to the inbox when
          a new message is received on this phone.
        example: false
        type: boolean
    required:
    - fcm_token
    - max_send_attempts
    - message_expiration_seconds
    - messages_per_minute
    - missed_call_auto_reply
    - phone_number
    - sim
    - unarchive_thread
    type: object
...
  responses.BadRequest:
    properties:
      data:
        example: The request body is not a valid JSON string
        type: string
      message:
        example: The request isn't properly formed
        type: string
      status:
        example: error
        type: string
    required:
    - data
    - message
    - status
    type: object
...
  responses.InternalServerError:
    properties:
      message:
        example: We ran into an internal error while handling the request.
        type: string
      status:
        example: error
        type: string
    required:
    - message
    - status
    type: object
...
  responses.NoContent:
    properties:
      message:
        example: action performed successfully
        type: string
      status:
        example: success
        type: string
    required:
    - message
    - status
    type: object
  responses.NotFound:
    properties:
      message:
        example: cannot find message with ID [32343a19-da5e-4b1b-a767-3298a73703ca]
        type: string
      status:
        example: error
        type: string
    required:
    - message
    - status
    type: object
...
  responses.PaymentRequired:
    properties:
      message:
        example: You have reached the maximum number of allowed resources. Please
          upgrade your plan.
        type: string
      status:
        example: error
        type: string
    required:
    - message
    - status
    type: object
  responses.PhoneAPIKeyResponse:
    properties:
      data:
        $ref: '#/definitions/entities.PhoneAPIKey'
      message:
        example: Request handled successfully
        type: string
      status:
        example: success
        type: string
    required:
    - data
    - message
    - status
    type: object
  responses.PhoneAPIKeysResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/entities.PhoneAPIKey'
        type: array
      message:
        example: Request handled successfully
        type: string
      status:
        example: success
        type: string
    required:
    - data
    - message
    - status
    type: object
  responses.PhoneResponse:
    properties:
      data:
        $ref: '#/definitions/entities.Phone'
      message:
        example: Request handled successfully
        type: string
      status:
        example: success
        type: string
    required:
    - data
    - message
    - status
    type: object
  responses.PhonesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/entities.Phone'
        type: array
      message:
        example: Request handled successfully
        type: string
      status:
        example: success
        type: string
    required:
    - data
    - message
    - status
    type: object
  responses.Unauthorized:
    properties:
      data:
        example: Make sure your API key is set in the [X-API-Key] header in the request
        type: string
      message:
        example: You are not authorized to carry out this request.
        type: string
      status:
        example: error
        type: string
    required:
    - data
    - message
    - status
    type: object
  responses.UnprocessableEntity:
    properties:
      data:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        example: validation errors while handling request
        type: string
      status:
        example: error
        type: string
    required:
    - data
    - message
    - status
    type: object
  responses.UserResponse:
    properties:
      data:
        $ref: '#/definitions/entities.User'
      message:
        example: Request handled successfully
        type: string
      status:
        example: success
        type: string
    required:
    - data
    - message
    - status
    type: object
...
host: api.httpsms.com
info:
  contact:
    email: support@httpsms.com
    name: support@httpsms.com
  description: Use your Android phone to send and receive SMS messages via a simple
    programmable API with end-to-end encryption.
  license:
    name: AGPL-3.0
    url: https://raw.githubusercontent.com/NdoleStudio/http-sms-manager/main/LICENSE
  title: httpSMS API Reference
  version: "1.0"
paths:
...
  /phone-api-keys:
    get:
      consumes:
      - application/json
      description: Get list phone API keys which a user has registered on the httpSMS
        application
      parameters:
      - description: number of phone api keys to skip
        in: query
        minimum: 0
        name: skip
        type: integer
      - description: filter phone api keys with name containing query
        in: query
        name: query
        type: string
      - description: number of phone api keys to return
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.PhoneAPIKeysResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Unauthorized'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.UnprocessableEntity'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerError'
      security:
      - ApiKeyAuth: []
      summary: Get the phone API keys of a user
      tags:
      - PhoneAPIKeys
    post:
      consumes:
      - application/json
      description: Creates a new phone API key which can be used to log in to the
        httpSMS app on your Android phone
      parameters:
      - description: Payload of new phone API key.
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.PhoneAPIKeyStoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.PhoneAPIKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Unauthorized'
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/responses.PaymentRequired'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.UnprocessableEntity'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerError'
      security:
      - ApiKeyAuth: []
      summary: Store phone API key
      tags:
      - PhoneAPIKeys
  /phone-api-keys/{phoneAPIKeyID}:
    delete:
      consumes:
      - application/json
      description: Delete a phone API Key from the database and cannot be used for
        authentication anymore.
      parameters:
      - default: 32343a19-da5e-4b1b-a767-3298a73703ca
        description: ID of the phone API key
        in: path
        name: phoneAPIKeyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/responses.NoContent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Unauthorized'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.NotFound'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.UnprocessableEntity'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerError'
      security:
      - ApiKeyAuth: []
      summary: Delete a phone API key from the database.
      tags:
      - PhoneAPIKeys
  /phone-api-keys/{phoneAPIKeyID}/phones/{phoneID}:
    delete:
      consumes:
      - application/json
      description: You will need to login again to the httpSMS app on your Android
        phone with a new phone API key.
      parameters:
      - default: 32343a19-da5e-4b1b-a767-3298a73703ca
        description: ID of the phone API key
        in: path
        name: phoneAPIKeyID
        required: true
        type: string
      - default: 32343a19-da5e-4b1b-a767-3298a73703ca
        description: ID of the phone
        in: path
        name: phoneID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/responses.NoContent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Unauthorized'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.NotFound'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.UnprocessableEntity'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerError'
      security:
      - ApiKeyAuth: []
      summary: Remove the association of a phone from the phone API key.
      tags:
      - PhoneAPIKeys
  /phones:
    get:
      consumes:
      - application/json
      description: Get list of phones which a user has registered on the http sms
        application
      parameters:
      - description: number of heartbeats to skip
        in: query
        minimum: 0
        name: skip
        type: integer
      - description: filter phones containing query
        in: query
        name: query
        type: string
      - description: number of phones to return
        in: query
        maximum: 20
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.PhonesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Unauthorized'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.UnprocessableEntity'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerError'
      security:
      - ApiKeyAuth: []
      summary: Get phones of a user
      tags:
      - Phones
    put:
      consumes:
      - application/json
      description: Updates properties of a user's phone. If the phone with this number
        does not exist, a new one will be created. Think of this method like an 'upsert'.
        URL-backed phone gateways receive FCM-compatible HTTP wake-ups.
      parameters:
      - description: Payload of new phone number.
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.PhoneUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.PhoneResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Unauthorized'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.UnprocessableEntity'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerError'
      security:
      - ApiKeyAuth: []
      summary: Upsert Phone
      tags:
      - Phones
  /phones/{phoneID}:
    delete:
      consumes:
      - application/json
      description: Delete a phone that has been sored in the database
      parameters:
      - default: 32343a19-da5e-4b1b-a767-3298a73703ca
        description: ID of the phone
        in: path
        name: phoneID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/responses.NoContent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Unauthorized'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.UnprocessableEntity'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerError'
      security:
      - ApiKeyAuth: []
      summary: Delete Phone
      tags:
      - Phones
...
  /users/{userID}/api-keys:
    delete:
      consumes:
      - application/json
      description: Rotate the user's API key in case the current API Key is compromised
      parameters:
      - default: 32343a19-da5e-4b1b-a767-3298a73703ca
        description: ID of the user to update
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Unauthorized'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.UnprocessableEntity'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerError'
      security:
      - ApiKeyAuth: []
      summary: Rotate the user's API Key
      tags:
      - Users
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-api-Key
    type: apiKey
swagger: "2.0"